requirements.txt
Flask==3.0.3
instagrapi==2.1.1
Pillow==10.3.0
google-generativeai==0.5.2
opencv-python==4.9.0.80
numpy==1.26.4